% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_outliers.R
\name{get_outliers}
\alias{get_outliers}
\title{Extract Outliers}
\usage{
get_outliers(data, fold_change)
}
\arguments{
\item{data}{A data.frame.}

\item{fold_change}{A numeric.}
}
\value{
A data.frame.
}
\description{
Extract Outliers
}
