% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_xlsx_contents.R
\name{get_xlsx_contents}
\alias{get_xlsx_contents}
\title{Read And Tidy Experiments Spreadsheets}
\usage{
get_xlsx_contents(
  files,
  project_name = NULL,
  od_outlier = 1.5,
  lm_outlier = 1.5
)
}
\arguments{
\item{files}{A character.}

\item{project_name}{A character.}

\item{od_outlier}{A numeric.}

\item{lm_outlier}{A numeric.}
}
\value{
A data.frame.
}
\description{
Read And Tidy Experiments Spreadsheets
}
