% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_outliers.R
\name{get_outliers}
\alias{get_outliers}
\title{Extract Outliers}
\usage{
get_outliers(data, fold_change)
}
\arguments{
\item{data}{A data.frame.
A \code{data.frame} from \code{get_xlsx_contents} in which
to extract measurements identified as outliers.}

\item{fold_change}{A numeric.
Fold-change threshold for reference above
which the reference is considered 'good'.}
}
\value{
A data.frame.
}
\description{
Extract Outliers
}
