% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_xlsx_contents.R
\name{get_xlsx_contents}
\alias{get_xlsx_contents}
\title{Read And Tidy Experiments Spreadsheets}
\usage{
get_xlsx_contents(
  files,
  project_name = NULL,
  od_outlier = 1.5,
  lm_outlier = 1.5
)
}
\arguments{
\item{files}{A character.
A path to an Excel file or a vector of paths to Excel files based on the template provided.
\code{system.file("app", "www", "template.xlsx", package = "insane")}.}

\item{project_name}{A character. A name under which the experiment or set of experiments
is to be stored within the shiny app}

\item{od_outlier}{A numeric. A multiplicator threshold based on Tukey's method
to define outliers in optical density measures.}

\item{lm_outlier}{A numeric. A multiplicator threshold based on Tukey's method
to define outliers in optical density measures for the blank intercept and slope estimates.}
}
\value{
A data.frame summarising all imported Excel files in addition to computed variables.
}
\description{
Read And Tidy Experiments Spreadsheets
}
