% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insane-package.R
\docType{package}
\name{insane-package}
\alias{insane}
\alias{insane-package}
\title{insane: INsulin Secretion ANalysEr}
\description{
A user-friendly interface, using Shiny, to analyse insulin secretion in cells.
    The Shiny App was initially designed for EndoC-betaH1 cell line following method described 
    in Ndiaye et al., 2017 (<doi:10.1016/j.molmet.2017.03.011>).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mcanouil/insane}
  \item \url{https://mcanouil.github.io/insane}
  \item Report bugs at \url{https://github.com/mcanouil/insane/issues}
}

}
\author{
\strong{Maintainer}: Mickaël Canouil \email{***REMOVED***} (\href{https://orcid.org/0000-0002-3396-4549}{ORCID})

}
\keyword{internal}
